+++ src/main/java/lotto/utils/WebUIRenderer.java
@@ -1,0 +1,12 @@
+package lotto.utils;
+
+import spark.ModelAndView;
+import spark.template.handlebars.HandlebarsTemplateEngine;
+
+import java.util.Map;
+
+public class WebUIRenderer {
+    public static String render(Map<String, Object> model, String templatePath) {
+        return new HandlebarsTemplateEngine().render(new ModelAndView(model, templatePath));
+    }
+}
